// src/app/core/api/generated/model/stock-response.ts
/**
 * Seggi App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StockResponse { 
    idStock?: number;
    productId?: number;
    productName?: string;
    warehouseId?: number;
    warehouseName?: string;
    quantity?: number;
}