// src/app/core/api/generated/model/warehouse-response.ts
/**
 * Seggi App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WarehouseResponse { 
    idWarehouse?: number;
    name?: string;
    address?: string;
    capacity?: number;
}